segregated api specfic code.
