+++ api/validators/team/validator.py
@@ -1,4 +1,4 @@
-from pyteam.mappings import validator_mappings
+from api.validators import mappings as validator_mappings
 
 
 class Validator(object):
@@ -6,18 +6,14 @@ class Validator(object):
     Class that wraps up all the methods necessary for validating arguments passed to TeamHandler.
     """
 
-    def __init__(self, handler, method, param_dict):
-        """
-        This method validates parameters passed to a handler based on action.
-        :param handler: name of handler
-        :param method: method name
-        :param param_dict: named args passed to method
-        :return: True if validates else raise exception.
-        """
-        validator_dict = validator_mappings.get(handler).get(method)
+    def get_validator_dict(self):
+        return validator_mappings.get(self)
+
+    def create_validator(self, **kwargs):
+        validator_dict = validator_mappings.get('TeamValidator').get('create')
         validated = False
         for key, value in validator_dict.items():
-            for param, param_value in param_dict.items():
+            for param, param_value in kwargs.items():
                 if key == param:
                     if type(value) is list:
                         if param_value in value:
@@ -29,5 +25,4 @@ class Validator(object):
                     else:
                         validated = False
         if not validated:
-            raise Exception('Validation Failed.')
-        return True
+            raise Exception('Validation error.')
